Apply REST error wrapper and split resource route tests
