+++ tests/api/test_routes.py
@@ -7,25 +7,8 @@ import httpx
 import pytest
 from httpx import ASGITransport, AsyncClient
 from mcp.server.fastmcp import FastMCP
-from mcp.types import Annotations, Resource
-from pydantic import AnyUrl
 
-from blockscout_mcp_server.api.routes import register_api_routes
 from blockscout_mcp_server.models import AdvancedFilterItem, TokenTransfer, ToolResponse, TransactionInfoData
-
-
-@pytest.fixture
-def test_mcp_instance():
-    """Provides a FastMCP instance for testing."""
-    return FastMCP(name="test-server-for-routes")
-
-
-@pytest.fixture
-def client(test_mcp_instance):
-    """Provides an httpx client configured to talk to the test MCP instance."""
-    register_api_routes(test_mcp_instance)
-    asgi_app = test_mcp_instance.streamable_http_app()
-    return AsyncClient(transport=ASGITransport(app=asgi_app), base_url="http://test")
 
 
 @pytest.mark.asyncio
@@ -72,46 +55,6 @@ async def test_list_tools_success(client: AsyncClient, test_mcp_instance: FastMC
     assert response.status_code == 200
     assert response.json() == []
     test_mcp_instance.list_tools.assert_called_once()
-
-
-@pytest.mark.asyncio
-async def test_list_resources_success(client: AsyncClient, test_mcp_instance: FastMCP):
-    """Verify that the /v1/resources endpoint returns a list of resources."""
-    test_mcp_instance.list_resources = AsyncMock(return_value=[])
-    response = await client.get("/v1/resources")
-    assert response.status_code == 200
-    assert response.json() == []
-    test_mcp_instance.list_resources.assert_called_once()
-
-
-@pytest.mark.asyncio
-async def test_list_resources_with_items(client: AsyncClient, test_mcp_instance: FastMCP):
-    """Verify resources are serialized with protocol aliases and JSON-safe URLs."""
-    resource = Resource(
-        uri=AnyUrl("blockscout-mcp://skill/SKILL.md"),
-        name="Blockscout Analysis Skill",
-        description="Bundled skill root file",
-        mimeType="text/markdown",
-        annotations=Annotations(audience=["assistant"], priority=0.9),
-        _meta={"source": "test"},
-    )
-
-    test_mcp_instance.list_resources = AsyncMock(return_value=[resource])
-
-    response = await client.get("/v1/resources")
-
-    assert response.status_code == 200
-    payload = response.json()
-    assert isinstance(payload, list)
-    assert len(payload) == 1
-    assert payload[0]["uri"] == "blockscout-mcp://skill/SKILL.md"
-    assert payload[0]["mimeType"] == "text/markdown"
-    assert payload[0]["annotations"]["audience"] == ["assistant"]
-    assert payload[0]["annotations"]["priority"] == 0.9
-    assert payload[0]["_meta"] == {"source": "test"}
-    assert "meta" not in payload[0]
-
-    test_mcp_instance.list_resources.assert_called_once()
 
 
 @pytest.mark.asyncio
